# Conflicts:
#	services/atlas-channel/atlas.com/channel/main.go
#	services/atlas-channel/atlas.com/channel/movement/processor_test.go
#	services/atlas-channel/atlas.com/channel/socket/handler/character_attack_drain_test.go

+++ services/atlas-query-aggregator/atlas.com/query-aggregator/character/rest.go
@@ -125,7 +125,7 @@ func Transform(m Model) (RestModel, error) {
 		Face:               m.Face(),
 		Ap:                 m.Ap(),
 		Sp:                 spStr,
-		SpawnPoint:         uint32(m.SpawnPoint()),
+		SpawnPoint:         m.SpawnPoint(),
 		Gm:                 m.GmLevel(),
 		X:                  m.X(),
 		Y:                  m.Y(),
